Figured out some kind of writeup to compare these different approaches to writing a custom hook
I want to also explore something like useReducer for useTimer.ts
Is it necessary to have things done in an fp approach from within the
hook? I want to say probably, or at the very least that that's what
React wants from us. Let me figure it out and see if I can write the
other approaches. If it works it works. If not, it's whatever.